finalización para sprint2

+++ server/database/migrations/2025_04_15_181848_create_nota_examen_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('examenId');
             $table->unsignedBigInteger('usuarioId');
-            $table->float('nota')->default(0.00);
+            $table->decimal('nota')->default(0.00);
             $table->timestamps();
 
             $table->foreign('examenId')->references('id')->on('examen')->onDelete('cascade');
